Renamed .Properties.Profile to Properties.ShapeProfiles and change of folder names
