+++ Structure_oM/Properties/Framing/ConstantFramingElementProperty.cs
