+++ Structure_oM/Properties/Framing/IFramingElementProperty.cs
