+++ Structure_oM/Properties/Section/CompositeSection.cs
