+++ Structure_oM/Properties/Section/ConcreteSection.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 using BH.oM.Common.Materials;
-using BH.oM.Structure.Properties.Section.Profiles;
+using BH.oM.Structure.Properties.Section.ShapeProfiles;
 
 namespace BH.oM.Structure.Properties.Section
 {
