+++ Structure_oM/Properties/Section/Enums/CableType.cs
