+++ Structure_oM/Properties/Section/Enums/ReinforcementPattern.cs
