+++ Structure_oM/Properties/Section/Enums/ShapeType.cs
