+++ Structure_oM/Properties/Section/Enums/SteelPlateRestraint.cs
