+++ Structure_oM/Properties/Section/ExplicitSection.cs
