+++ Structure_oM/Properties/Section/ISectionProperty.cs
