+++ Structure_oM/Properties/Section/ShapeProfiles/IProfile.cs
@@ -2,7 +2,7 @@
 using BH.oM.Geometry;
 using System.Collections.ObjectModel;
 
-namespace BH.oM.Structure.Properties.Section.Profiles
+namespace BH.oM.Structure.Properties.Section.ShapeProfiles
 {
     public interface IProfile  : IBHoMObject
     {
