+++ Structure_oM/Properties/Section/ShapeProfiles/KiteProfile.cs
@@ -1,19 +1,24 @@
-﻿using BH.oM.Base;
-using BH.oM.Geometry;
-using System.Collections.ObjectModel;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
+using System.Collections.ObjectModel;
+using BH.oM.Base;
+using BH.oM.Geometry;
 
-namespace BH.oM.Structure.Properties.Section.Profiles
+namespace BH.oM.Structure.Properties.Section.ShapeProfiles
 {
-    public class CircleProfile : BHoMObject, IProfile, IImmutable
+    public class KiteProfile : BHoMObject, IProfile, IImmutable
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
-        public ShapeType Shape { get; } = ShapeType.Circle;
 
-        public double Diameter { get; }
+        public ShapeType Shape { get; } = ShapeType.DoubleAngle;
+
+        public double Width1 { get; }
+
+        public double Angle1 { get; } 
+
+        public double Thickness { get; }
 
         public ReadOnlyCollection<ICurve> Edges { get; }
 
@@ -21,12 +26,13 @@ namespace BH.oM.Structure.Properties.Section.Profiles
         /**** Constructors                              ****/
         /***************************************************/
 
-        public CircleProfile(double diameter, IEnumerable<ICurve> edges)
+        public KiteProfile(double width1, double angle1, double thickness, IEnumerable<ICurve> edges)
         {
-            Diameter = diameter;
+            Width1 = width1;
+            Angle1 = angle1;
+            Thickness = thickness;
             Edges = new ReadOnlyCollection<ICurve>(edges.ToList());
         }
-
 
         /***************************************************/
     }
