+++ Structure_oM/Properties/Section/ShapeProfiles/TSectionProfile.cs
@@ -4,15 +4,15 @@ using System.Collections.ObjectModel;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace BH.oM.Structure.Properties.Section.Profiles
+namespace BH.oM.Structure.Properties.Section.ShapeProfiles
 {
-    public class AngleProfile : BHoMObject, IProfile, IImmutable
+    public class TSectionProfile : BHoMObject, IProfile, IImmutable
     {
 
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
-        public ShapeType Shape { get; } = ShapeType.Angle;
+        public ShapeType Shape { get; } = ShapeType.Tee;
 
         public double Height { get; }
 
@@ -26,18 +26,15 @@ namespace BH.oM.Structure.Properties.Section.Profiles
 
         public double ToeRadius { get; }
 
-        public bool MirrorAboutLocalZ { get; }
-
         public bool MirrorAboutLocalY { get; }
 
         public ReadOnlyCollection<ICurve> Edges { get; }
-
 
         /***************************************************/
         /**** Constructors                              ****/
         /***************************************************/
 
-        public AngleProfile(double height, double width, double webthickness, double flangeThickness, double rootRadius, double toeRadius, bool mirrorAboutLocalZ, bool mirrorAboutLocalY, IEnumerable<ICurve> edges)
+        public TSectionProfile(double height, double width, double webthickness, double flangeThickness, double rootRadius, double toeRadius, bool mirrorAboutLocalY, IEnumerable<ICurve> edges)
         {
             Height = height;
             Width = width;
@@ -45,7 +42,6 @@ namespace BH.oM.Structure.Properties.Section.Profiles
             FlangeThickness = flangeThickness;
             RootRadius = rootRadius;
             ToeRadius = toeRadius;
-            MirrorAboutLocalZ = mirrorAboutLocalZ;
             MirrorAboutLocalY = mirrorAboutLocalY;
             Edges = new ReadOnlyCollection<ICurve>(edges.ToList());
         }
