+++ Structure_oM/Properties/Section/ShapeProfiles/TubeProfile.cs
@@ -4,14 +4,19 @@ using System.Collections.ObjectModel;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace BH.oM.Structure.Properties.Section.Profiles
+namespace BH.oM.Structure.Properties.Section.ShapeProfiles
 {
-    public class FreeFormProfile : BHoMObject, IProfile, IImmutable
+    public class TubeProfile : BHoMObject, IProfile, IImmutable
     {
+
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
-        public ShapeType Shape { get; } = ShapeType.FreeForm;
+        public ShapeType Shape { get; } = ShapeType.Tube;
+
+        public double Diameter { get; }
+
+        public double Thickness { get; }
 
         public ReadOnlyCollection<ICurve> Edges { get; }
 
@@ -19,8 +24,10 @@ namespace BH.oM.Structure.Properties.Section.Profiles
         /**** Constructors                              ****/
         /***************************************************/
 
-        public FreeFormProfile(IEnumerable<ICurve> edges)
+        public TubeProfile(double diameter, double thickness, IEnumerable<ICurve> edges)
         {
+            Diameter = diameter;
+            Thickness = thickness;
             Edges = new ReadOnlyCollection<ICurve>(edges.ToList());
         }
