+++ Structure_oM/Properties/Section/ShapeProfiles/ZSectionProfile.cs
@@ -4,7 +4,7 @@ using System.Collections.ObjectModel;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace BH.oM.Structure.Properties.Section.Profiles
+namespace BH.oM.Structure.Properties.Section.ShapeProfiles
 {
     public class ZSectionProfile : BHoMObject, IProfile, IImmutable
     {
