+++ Structure_oM/Properties/Section/SteelSection.cs
@@ -1,6 +1,6 @@
 ﻿using BH.oM.Base;
 using BH.oM.Common.Materials;
-using BH.oM.Structure.Properties.Section.Profiles;
+using BH.oM.Structure.Properties.Section.ShapeProfiles;
 
 namespace BH.oM.Structure.Properties.Section
 {
