+++ Structure_oM/Properties/Surface/ConcreteWaffle.cs
