+++ Structure_oM/Properties/Surface/ConstantThickness.cs
