+++ Structure_oM/Properties/Surface/Enums/LoadPanelSupportConditions.cs
