+++ Structure_oM/Properties/Surface/Enums/PanelDirection.cs
