+++ Structure_oM/Properties/Surface/Enums/PanelType.cs
