+++ Structure_oM/Structure_oM.csproj
@@ -88,50 +88,50 @@
     <Compile Include="Properties\Constraints\Constraint4DOF.cs" />
     <Compile Include="Properties\Constraints\Constraint6DOF.cs" />
     <Compile Include="Properties\Constraints\LinkConstraint.cs" />
-    <Compile Include="Properties\Section Properties\Enums\CableType.cs" />
+    <Compile Include="Properties\Section\Enums\CableType.cs" />
     <Compile Include="Properties\Constraints\Enums\DOFType.cs" />
-    <Compile Include="Properties\Property2D\Enums\LoadPanelSupportConditions.cs" />
-    <Compile Include="Properties\Property2D\Enums\PanelDirection.cs" />
-    <Compile Include="Properties\Property2D\Enums\PanelModifier.cs" />
-    <Compile Include="Properties\Property2D\Enums\PanelType.cs" />
-    <Compile Include="Properties\Section Properties\Enums\ReinforcementPattern.cs" />
-    <Compile Include="Properties\Section Properties\Enums\ShapeType.cs" />
-    <Compile Include="Properties\Section Properties\Enums\SteelFabrication.cs" />
-    <Compile Include="Properties\Section Properties\Enums\SteelPlateRestraint.cs" />
-    <Compile Include="Properties\FramingElementProperty\ConstantFramingElementProperty.cs" />
-    <Compile Include="Properties\FramingElementProperty\IFramingElementProperty.cs" />
+    <Compile Include="Properties\Surface\Enums\LoadPanelSupportConditions.cs" />
+    <Compile Include="Properties\Surface\Enums\PanelDirection.cs" />
+    <Compile Include="Properties\Surface\Enums\PanelModifier.cs" />
+    <Compile Include="Properties\Surface\Enums\PanelType.cs" />
+    <Compile Include="Properties\Section\Enums\ReinforcementPattern.cs" />
+    <Compile Include="Properties\Section\Enums\ShapeType.cs" />
+    <Compile Include="Properties\Section\Enums\SteelFabrication.cs" />
+    <Compile Include="Properties\Section\Enums\SteelPlateRestraint.cs" />
+    <Compile Include="Properties\Framing\ConstantFramingElementProperty.cs" />
+    <Compile Include="Properties\Framing\IFramingElementProperty.cs" />
     <Compile Include="Properties\Offset.cs" />
-    <Compile Include="Properties\Property2D\ConcreteRibbed.cs" />
-    <Compile Include="Properties\Property2D\ConcreteWaffle.cs" />
-    <Compile Include="Properties\Property2D\ConstantThickness.cs" />
-    <Compile Include="Properties\Property2D\ISurfaceProperty.cs" />
-    <Compile Include="Properties\Property2D\LoadingPanel.cs" />
-    <Compile Include="Properties\Section Properties\CableSection.cs" />
-    <Compile Include="Properties\Section Properties\CompositeSection.cs" />
-    <Compile Include="Properties\Section Properties\ConcreteSection.cs" />
-    <Compile Include="Properties\Section Properties\ExplicitSection.cs" />
-    <Compile Include="Properties\Section Properties\ISectionProperty.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\AngleProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\BoxProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\ChannelProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\CircleProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\FabricatedBoxProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\FabricatedISectionProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\FreeFormProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\GeneralisedFabricatedBoxProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\GeneralisedTSectionProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\IProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\ISectionProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\KiteProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\RectangleProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\TSectionProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\TubeProfile.cs" />
-    <Compile Include="Properties\Section Properties\Profiles\ZSectionProfile.cs" />
-    <Compile Include="Properties\Section Properties\Reinforcement\LayerReinforcement.cs" />
-    <Compile Include="Properties\Section Properties\Reinforcement\PerimeterReinforcement.cs" />
-    <Compile Include="Properties\Section Properties\Reinforcement\Reinforcement.cs" />
-    <Compile Include="Properties\Section Properties\Reinforcement\TieReinforcement.cs" />
-    <Compile Include="Properties\Section Properties\SteelSection.cs" />
+    <Compile Include="Properties\Surface\ConcreteRibbed.cs" />
+    <Compile Include="Properties\Surface\ConcreteWaffle.cs" />
+    <Compile Include="Properties\Surface\ConstantThickness.cs" />
+    <Compile Include="Properties\Surface\ISurfaceProperty.cs" />
+    <Compile Include="Properties\Surface\LoadingPanel.cs" />
+    <Compile Include="Properties\Section\CableSection.cs" />
+    <Compile Include="Properties\Section\CompositeSection.cs" />
+    <Compile Include="Properties\Section\ConcreteSection.cs" />
+    <Compile Include="Properties\Section\ExplicitSection.cs" />
+    <Compile Include="Properties\Section\ISectionProperty.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\AngleProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\BoxProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\ChannelProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\CircleProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\FabricatedBoxProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\FabricatedISectionProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\FreeFormProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\GeneralisedFabricatedBoxProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\GeneralisedTSectionProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\IProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\ISectionProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\KiteProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\RectangleProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\TSectionProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\TubeProfile.cs" />
+    <Compile Include="Properties\Section\ShapeProfiles\ZSectionProfile.cs" />
+    <Compile Include="Properties\Section\Reinforcement\LayerReinforcement.cs" />
+    <Compile Include="Properties\Section\Reinforcement\PerimeterReinforcement.cs" />
+    <Compile Include="Properties\Section\Reinforcement\Reinforcement.cs" />
+    <Compile Include="Properties\Section\Reinforcement\TieReinforcement.cs" />
+    <Compile Include="Properties\Section\SteelSection.cs" />
     <Compile Include="Results\Bar Results\BarDeformation.cs" />
     <Compile Include="Results\Bar Results\BarForce.cs" />
     <Compile Include="Results\Bar Results\BarResult.cs" />
